Moved things around to better places
